Eighth Commit (Requirements.md) 10/23 - Added artist removal feature with albums and songs only if this artist is exclusive owner of it, if it's featured with another artist, deleting only relation and artist from db leaving songs and albums untouched.

+++ src/main/java/com/songify/domain/crud/AlbumDeleter.java
@@ -14,7 +14,7 @@ class AlbumDeleter {
 
     void deleteById(Album album) {
         if (album.getSongs().isEmpty()) {
-            albumRepository.deleteById(album.getId());
+            albumRepository.deleteById(album.getId().longValue());
         } else {
             throw new AlbumNotEmptyException("Album is not empty, cannot delete");
         }
